requirements.txt
django==1.11.1
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
coreapi==2.3.0
coreapi-cli==1.0.6
django-cors-headers==2.0.2
